Rename/reorg e2e helpers (#5127)
